fix(admin): share formatRelative/formatDateTime via format module
UserAnalyticsPanel referenced formatRelative without importing it
(it was a local non-exported helper in UserAnalyticsSections),
breaking the frontend build. Move both helpers into the shared
user-analytics-format module and import them in both components.
